add video to template detail page
