+++ src/components/TemplateDetailPage/TemplateDetailPage.css
@@ -10,12 +10,25 @@
   margin-right: 8px;
 }
 
-.TemplateDetailPage .header-image {
+.TemplateDetailPage .template-media {
   height: 347px;
   width: 100%;
-  background-color: var(--card);
+  position: relative;
+}
+
+.TemplateDetailPage .header-image {
+  display: block;
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  object-fit: cover;
   background-size: cover;
   background-position: center;
+  object-fit: cover;
+}
+
+.TemplateDetailPage .header-image.image-hidden {
+  display: none;
 }
 
 .TemplateDetailPage .info-container {
